separating rest controller interfaces
