+++ activiti-services/activiti-services-rest/src/main/java/org/activiti/services/rest/resources/assembler/ProcessInstanceResourceAssembler.java
@@ -18,8 +18,8 @@ package org.activiti.services.rest.resources.assembler;
 import org.activiti.services.core.model.ProcessInstance;
 import org.activiti.services.rest.resources.ProcessInstanceResource;
 import org.activiti.services.rest.controllers.HomeController;
-import org.activiti.services.rest.controllers.ProcessInstanceController;
-import org.activiti.services.rest.controllers.ProcessInstanceVariableController;
+import org.activiti.services.rest.controllers.ProcessInstanceControllerImpl;
+import org.activiti.services.rest.controllers.ProcessInstanceVariableControllerImpl;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.mvc.ResourceAssemblerSupport;
 import org.springframework.stereotype.Component;
@@ -31,15 +31,15 @@ import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
 public class ProcessInstanceResourceAssembler extends ResourceAssemblerSupport<ProcessInstance, ProcessInstanceResource> {
 
     public ProcessInstanceResourceAssembler() {
-        super(ProcessInstanceController.class, ProcessInstanceResource.class);
+        super(ProcessInstanceControllerImpl.class, ProcessInstanceResource.class);
     }
 
     @Override
     public ProcessInstanceResource toResource(ProcessInstance processInstance) {
-        Link processInstancesRel = linkTo(methodOn(ProcessInstanceController.class).getProcessInstances(null, null))
+        Link processInstancesRel = linkTo(methodOn(ProcessInstanceControllerImpl.class).getProcessInstances(null, null))
                                                                                                                     .withRel("processInstances");
-        Link selfLink = linkTo(methodOn(ProcessInstanceController.class).getProcessInstanceById(processInstance.getId())).withSelfRel();
-        Link variablesLink = linkTo(methodOn(ProcessInstanceVariableController.class).getVariables(processInstance.getId())).withRel("variables");
+        Link selfLink = linkTo(methodOn(ProcessInstanceControllerImpl.class).getProcessInstanceById(processInstance.getId())).withSelfRel();
+        Link variablesLink = linkTo(methodOn(ProcessInstanceVariableControllerImpl.class).getVariables(processInstance.getId())).withRel("variables");
         Link homeLink = linkTo(HomeController.class).withRel("home");
         return new ProcessInstanceResource(processInstance, selfLink, variablesLink, processInstancesRel, homeLink);
     }
